feat: crate app-context
