+++ .eslintcache
@@ -1,0 +1,1 @@
+[{"C:\\Users\\ikryu\\Desktop\\моя работа\\React-app\\src\\components\\UI\\modal.tsx":"1","C:\\Users\\ikryu\\Desktop\\моя работа\\React-app\\src\\components\\pagination.tsx":"2","C:\\Users\\ikryu\\Desktop\\моя работа\\React-app\\src\\components\\resultList.tsx":"3","C:\\Users\\ikryu\\Desktop\\моя работа\\React-app\\src\\components\\router\\routing.tsx":"4","C:\\Users\\ikryu\\Desktop\\моя работа\\React-app\\src\\interface\\interface.tsx":"5","C:\\Users\\ikryu\\Desktop\\моя работа\\React-app\\src\\pages\\detailedPage.tsx":"6","C:\\Users\\ikryu\\Desktop\\моя работа\\React-app\\src\\pages\\searchPage.tsx":"7"},{"size":247,"mtime":1699557842881,"results":"8","hashOfConfig":"9"},{"size":4114,"mtime":1699561534249,"results":"10","hashOfConfig":"9"},{"size":479,"mtime":1699558807273,"results":"11","hashOfConfig":"9"},{"size":370,"mtime":1699557870278,"results":"12","hashOfConfig":"9"},{"size":2252,"mtime":1699558694901,"results":"13","hashOfConfig":"9"},{"size":564,"mtime":1699561442319,"results":"14","hashOfConfig":"9"},{"size":7544,"mtime":1699561629327,"results":"15","hashOfConfig":"9"},{"filePath":"16","messages":"17","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"1cl95ay",{"filePath":"18","messages":"19","errorCount":0,"fatalErrorCount":0,"warningCount":1,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},{"filePath":"20","messages":"21","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"22","messages":"23","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"24","messages":"25","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"26","messages":"27","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"28","messages":"29","errorCount":0,"fatalErrorCount":0,"warningCount":2,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},"C:\\Users\\ikryu\\Desktop\\моя работа\\React-app\\src\\components\\UI\\modal.tsx",[],"C:\\Users\\ikryu\\Desktop\\моя работа\\React-app\\src\\components\\pagination.tsx",["30"],"C:\\Users\\ikryu\\Desktop\\моя работа\\React-app\\src\\components\\resultList.tsx",[],"C:\\Users\\ikryu\\Desktop\\моя работа\\React-app\\src\\components\\router\\routing.tsx",[],"C:\\Users\\ikryu\\Desktop\\моя работа\\React-app\\src\\interface\\interface.tsx",[],"C:\\Users\\ikryu\\Desktop\\моя работа\\React-app\\src\\pages\\detailedPage.tsx",[],"C:\\Users\\ikryu\\Desktop\\моя работа\\React-app\\src\\pages\\searchPage.tsx",["31","32"],{"ruleId":"33","severity":1,"message":"34","line":34,"column":6,"nodeType":"35","endLine":34,"endColumn":18,"suggestions":"36"},{"ruleId":"33","severity":1,"message":"37","line":40,"column":6,"nodeType":"35","endLine":40,"endColumn":8,"suggestions":"38"},{"ruleId":"33","severity":1,"message":"39","line":81,"column":6,"nodeType":"35","endLine":81,"endColumn":67,"suggestions":"40"},"react-hooks/exhaustive-deps","React Hook useEffect has a missing dependency: 'initialButtonsState'. Either include it or remove the dependency array.","ArrayExpression",["41"],"React Hook useEffect has missing dependencies: 'handleQueryChange', 'queryParams', and 'searchState.page'. Either include them or remove the dependency array.",["42"],"React Hook useEffect has a missing dependency: 'fetchData'. Either include it or remove the dependency array.",["43"],{"desc":"44","fix":"45"},{"desc":"46","fix":"47"},{"desc":"48","fix":"49"},"Update the dependencies array to be: [initialButtonsState, totalPages]",{"range":"50","text":"51"},"Update the dependencies array to be: [handleQueryChange, queryParams, searchState.page]",{"range":"52","text":"53"},"Update the dependencies array to be: [searchState.limit, searchState.queryParam, searchState.page, fetchData]",{"range":"54","text":"55"},[1040,1052],"[initialButtonsState, totalPages]",[1508,1510],"[handleQueryChange, queryParams, searchState.page]",[2658,2719],"[searchState.limit, searchState.queryParam, searchState.page, fetchData]"]
