+++ src/App.tsx
@@ -1,8 +1,13 @@
 import { RouterProvider } from 'react-router-dom';
 import { router } from './components/router/routing';
+import { SearchProvider } from './contexts/SearchContext';
 
 function App() {
-  return <RouterProvider router={router} />;
+  return (
+    <SearchProvider>
+      <RouterProvider router={router} />;
+    </SearchProvider>
+  );
 }
 
 export default App;
